requirements.txt
pandas
openpyxl
pyyaml
requests
xsdata
linkml-runtime
xmlschema 